chore: clean up webpack config files for docs addon
CACTUS-1087
